Remove ExceptionUtils to fix license issues
