+++ LICENSE
@@ -211,4 +211,3 @@ The text of each license is the standard Apache 2.0 license.
 
 mvnw files from https://github.com/apache/maven-wrapper Apache 2.0
 streampark-console/streampark-console-service/src/main/assembly/bin/setclasspath.sh from https://github.com/apache/tomcat
-streampark-common/src/main/scala/org/apache/streampark/common/util/ExceptionUtils.scala from https://github.com/apache/flink
